replication impl complete not tested

+++ map-api/include/map-api/net-table-index.h
@@ -82,12 +82,13 @@ class NetTableIndex : public ChordIndex {
       const PeerId& to, DataMap* responsibilities) final override;
   virtual bool pushResponsibilitiesRpc(
       const PeerId& to, const DataMap& responsibilities) final override;
-  virtual bool initReplicatorRpc(
-      const PeerId& to, int index, const DataMap& data) final override;
-  virtual bool appendOnReplicatorRpc(
-      const PeerId& to, int index, const DataMap& data) final override;
-  virtual bool fetchFromReplicatorRpc(
-      const PeerId& to, int index, DataMap* data, PeerId* peer) final override;
+  virtual bool initReplicatorRpc(const PeerId& to, size_t index,
+                                 const DataMap& data) final override;
+  virtual bool appendOnReplicatorRpc(const PeerId& to, size_t index,
+                                     const DataMap& data) final override;
+  virtual bool fetchFromReplicatorRpc(const PeerId& to, size_t index,
+                                      DataMap* data,
+                                      PeerId* peer) final override;
 
   std::string table_name_;
   PeerHandler peers_;
