+++ map-api/include/map-api/spatial-index.h
@@ -185,12 +185,13 @@ class SpatialIndex : public ChordIndex {
       const PeerId& to, DataMap* responsibilities) final override;
   virtual bool pushResponsibilitiesRpc(
       const PeerId& to, const DataMap& responsibilities) final override;
-  virtual bool initReplicatorRpc(
-      const PeerId& to, int index, const DataMap& data) final override;
-  virtual bool appendOnReplicatorRpc(
-      const PeerId& to, int index, const DataMap& data) final override;
-  virtual bool fetchFromReplicatorRpc(
-      const PeerId& to, int index, DataMap* data, PeerId* peer) final override;
+  virtual bool initReplicatorRpc(const PeerId& to, size_t index,
+                                 const DataMap& data) final override;
+  virtual bool appendOnReplicatorRpc(const PeerId& to, size_t index,
+                                     const DataMap& data) final override;
+  virtual bool fetchFromReplicatorRpc(const PeerId& to, size_t index,
+                                      DataMap* data,
+                                      PeerId* peer) final override;
 
   virtual void localUpdateCallback(const std::string& key,
                                    const std::string& old_value,
