+++ map-api/src/net-table-index.cc
@@ -270,7 +270,7 @@ void NetTableIndex::handleRoutedRequest(
     }
     return;
   }
-  
+
   if (request.isType<kInitReplicatorRequest>()) {
     DataMap data;
     proto::FetchResponsibilitiesResponse init_request;
@@ -286,7 +286,7 @@ void NetTableIndex::handleRoutedRequest(
     }
     return;
   }
-  
+
   if (request.isType<kAppendReplicationDataRequest>()) {
     DataMap data;
     proto::FetchResponsibilitiesResponse replication_request;
@@ -302,7 +302,7 @@ void NetTableIndex::handleRoutedRequest(
     }
     return;
   }
-  
+
   if (request.isType<kFetchReplicationDataRequest>()) {
     DataMap data;
     PeerId replicating_peer_;
@@ -492,8 +492,8 @@ bool NetTableIndex::pushResponsibilitiesRpc(
   return true;
 }
 
-bool NetTableIndex::initReplicatorRpc(
-    const PeerId& to, int index, const DataMap& data) {
+bool NetTableIndex::initReplicatorRpc(const PeerId& to, size_t index,
+                                      const DataMap& data) {
   Message request, response;
   proto::FetchResponsibilitiesResponse push_request;
   for (const DataMap::value_type& item : data) {
@@ -510,8 +510,8 @@ bool NetTableIndex::initReplicatorRpc(
   return true;
 }
 
-bool NetTableIndex::appendOnReplicatorRpc(
-    const PeerId& to, int index, const DataMap& data) {
+bool NetTableIndex::appendOnReplicatorRpc(const PeerId& to, size_t index,
+                                          const DataMap& data) {
   Message request, response;
   proto::FetchResponsibilitiesResponse push_request;
   for (const DataMap::value_type& item : data) {
@@ -528,8 +528,8 @@ bool NetTableIndex::appendOnReplicatorRpc(
   return true;
 }
 
-bool NetTableIndex::fetchFromReplicatorRpc(
-    const PeerId& to, int index, DataMap* data, PeerId* peer) {
+bool NetTableIndex::fetchFromReplicatorRpc(const PeerId& to, size_t index,
+                                           DataMap* data, PeerId* peer) {
   CHECK_NOTNULL(data);
   CHECK_NOTNULL(peer);
   Message request, response;
